fix query for modes, add reports
